requirements.txt
alembic==1.4.2
blinker==1.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Cors==3.0.9
Flask-JWT-Extended==3.24.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
mypy==0.782
mypy-extensions==0.4.3
openpyxl==3.0.5
passlib==1.7.2
psycopg2-binary==2.8.6
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
redis==3.5.3
requests==2.24.0
rq==1.5.2
six==1.15.0
SQLAlchemy==1.3.19
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.11
Werkzeug==1.0.1
